Refactored Gene and Interval classes into their own files
