+++ src/transcriptmapping.cpp
@@ -4,107 +4,6 @@
 using std::string;
 using namespace Rcpp;
 
-Gene::Gene(string id, int st, int en, int snd): Interval(st, en, snd), gene_id(id) {}
-Gene::Gene(string id, int snd): Interval(-1, -1, snd), gene_id(id) {}
-Gene::Gene(): Interval(-1, -1, 0), gene_id("") {}
-
-void Gene::set_ID(string id)
-{
-    gene_id = id;
-}
-
-void Gene::add_exon(Interval it)
-{
-        exon_vec.push_back(it);
-        // expand the gene interval to include the new exon
-        if (st>it.st || st<0)
-        {
-            st = it.st;
-        }
-        if (en < it.en || en<0)
-        {
-            en = it.en;
-        }
-        if (snd == 0)
-        {
-            snd = it.snd;
-        }
-}
-
-int Gene::distance_to_end(Interval it)
-{
-    int distance = 0;
-    int tmp_en = 0;
-    auto iter = std::lower_bound(exon_vec.begin(), exon_vec.end(), it);
-    if (snd == 1)
-    {
-        distance += iter->en - ((iter->st)>it.st?(iter->st):it.st);
-        tmp_en = iter->en;
-        for (auto i = iter+1; i != exon_vec.end(); ++i)
-        {
-            if (tmp_en < i->st)
-            {
-                distance += i->en - i->st;
-                tmp_en = i->en;
-            }
-
-        }
-
-    }
-    else if (snd == -1)
-    {
-        for (auto i = exon_vec.begin(); i != iter; ++i)
-        {
-            if (tmp_en < i->st)
-            {
-                distance += i->en - i->st;
-                tmp_en = i->en;
-            }
-        }
-        if (tmp_en < iter->st)
-        {
-            distance += ((iter->en)<it.en?(iter->en):it.en) - iter->st;
-        }
-    }
-
-    return distance;
-}
-
-bool Gene::in_exon(Interval it)
-{
-    return std::binary_search(exon_vec.begin(), exon_vec.end(), it);
-}
-
-bool Gene::in_exon(Interval it, bool check_strand)
-{
-    if (check_strand && (it.snd*snd == -1))
-    {
-        return false;
-    }
-    else
-    {
-        return std::binary_search(exon_vec.begin(), exon_vec.end(), it);
-    }
-}
-
-void Gene::sort_exon()
-{
-    std::sort(exon_vec.begin(), exon_vec.end());
-}
-
-std::ostream& operator<< (std::ostream& out, const Gene& obj)
-{
-    out << "Gene ID:   " << obj.gene_id  << "\n";
-    out << "\t" << "start/end:   " << obj.st  << "/" << obj.en << "\n";
-    out << "\t" << "strand:   " << obj.snd  << "\n";
-    out << "\t" << "number of exons:   " << obj.exon_vec.size()  << "\n";
-    for (int i = 0; i < obj.exon_vec.size(); ++i)
-    {
-        out << "\t" << "exon[" << i+1 << "]: (" << obj.exon_vec[i].st << ", " << obj.exon_vec[i].en << ")" << "\n";
-    }
-    return out;
-}
-
 int GeneAnnotation::get_strand(char st)
 {
     int strand = 0;
@@ -369,7 +268,8 @@ int Mapping::map_exon(bam_hdr_t *header, bam1_t *b, string& gene_id, bool m_stra
         if (((bam_cigar_type(cig[c]) >> 0) & 1) && ((bam_cigar_type(cig[c]) >> 1) & 1))
         {
             Interval it = Interval(tmp_pos, tmp_pos+bam_cigar_oplen(cig[c]), rev);
-            auto iter = std::equal_range(Anno.gene_dict[header->target_name[b->core.tid]].begin(), Anno.gene_dict[header->target_name[b->core.tid]].end(), it);
+            auto gene_list = Anno.gene_dict[header->target_name[b->core.tid]];
+            auto iter = std::equal_range(gene_list.begin(), gene_list.end(), it);
             if ((iter.second - iter.first) == 0)
             {
                 tmp_ret = tmp_ret>3?3:tmp_ret;
