+++ src/transcriptmapping.h
@@ -10,36 +10,11 @@
 #include <Rcpp.h>
 #include "config_hts.h"
 #include "utils.h"
+#include "Gene.h"
+#include "Interval.h"
 
 #ifndef TRANSCRIPTMAPPING_H
 #define TRANSCRIPTMAPPING_H
-
-
-// for position, we dont store chromosome info because we will put genes from the same chromosome
-// together in a hashmap
-class Gene: public Interval
-{
-public:
-    std::string gene_id;
-    std::vector<Interval> exon_vec;
-
-    Gene(std::string id, int st, int en, int snd);
-    Gene(std::string id, int snd);
-    Gene();
-
-    void set_ID(std::string id);
-
-    void add_exon(Interval it);
-
-    bool in_exon(Interval it);
-    bool in_exon(Interval it, bool check_strand);
-
-    int distance_to_end(Interval it);
-    void sort_exon();
-
-    friend std::ostream& operator<< (std::ostream& out, const Gene& obj);
-    
-};
 
 // parse gff3 genome annotation
 class GeneAnnotation
