+++ src/utils.cpp
@@ -53,36 +53,6 @@ std::unordered_map<std::string, int> vector_counter(std::vector<std::string> v)
     return counter;
 }
 
-// Constructors for Interval
-Interval::Interval(int s, int e): st(s), en(e), snd(0) {}
-Interval::Interval(int s, int e, int sd): st(s), en(e), snd(sd) {}
-
-int Interval::overlap(int st1, int en1)
-{
-  if (en1 < st)
-  {
-    return -1;
-  }
-  else if (st1 > en)
-  {
-    return 1;
-  }
-  else
-  {
-    return 0;
-  }
-}
-
-bool operator < (const Interval &L, const Interval &R)
-{
-  return L.en < R.st;
-}
-
-bool operator > (const Interval &L, const Interval &R)
-{
-  return L.st > R.en;
-}
-
 std::vector<std::string> &split(const std::string &s, char delim, std::vector<std::string> &elems)
 {
     std::stringstream ss(s);
